Refactor mutex package and improve logging
Move signal constants and Handler struct to mutex.go. Remove unused
types.go file. Update MutexWorkflow to use a single periodic timer for
inactivity. Simplify logging messages in MutexWorkflow. Use maps.Copy
for copying hints in QuantmError.
